requirements.txt
Flask==3.1.2
zeroconf==0.148.0
flask_sqlalchemy==3.1.1